finally added some property based tests. Fin

+++ roman-numerals/roman_numerals.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 )
 
-func ConvertToRoman(arabic int) string {
+func ConvertToRoman(arabic uint16) string {
 	var result strings.Builder
 
 	for _, numeral := range allRomanNumerals {
@@ -17,7 +17,7 @@ func ConvertToRoman(arabic int) string {
 	return result.String()
 }
 
-func ConvertToArabic(roman string) (total int) {
+func ConvertToArabic(roman string) (total uint16) {
 	for _, symbol := range windowedRoman(roman).Symbols() {
 		total += allRomanNumerals.ValueOf(symbol...)
 	}
@@ -25,13 +25,13 @@ func ConvertToArabic(roman string) (total int) {
 }
 
 type romanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
 type romanNumerals []romanNumeral
 
-func (r romanNumerals) ValueOf(symbols ...byte) int {
+func (r romanNumerals) ValueOf(symbols ...byte) uint16 {
 	symbol := string(symbols)
 	for _, s := range r {
 		if s.Symbol == symbol {
